+++ src/restic/snapshot_find.go
@@ -12,7 +12,7 @@ import (
 var ErrNoSnapshotFound = errors.New("no snapshot found")
 
 // FindLatestSnapshot finds latest snapshot with optional target/directory, tags and hostname filters.
-func FindLatestSnapshot(ctx context.Context, repo Repository, targets []string, tags []string, hostname string) (ID, error) {
+func FindLatestSnapshot(ctx context.Context, repo Repository, targets []string, tagLists []TagList, hostname string) (ID, error) {
 	var (
 		latest   time.Time
 		latestID ID
@@ -24,11 +24,21 @@ func FindLatestSnapshot(ctx context.Context, repo Repository, targets []string, 
 		if err != nil {
 			return ID{}, errors.Errorf("Error listing snapshot: %v", err)
 		}
-		if snapshot.Time.After(latest) && (hostname == "" || hostname == snapshot.Hostname) && snapshot.HasTags(tags) && snapshot.HasPaths(targets) {
-			latest = snapshot.Time
-			latestID = snapshotID
-			found = true
+		if snapshot.Time.Before(latest) || (hostname != "" && hostname != snapshot.Hostname) {
+			continue
 		}
+
+		if !snapshot.HasTagList(tagLists) {
+			continue
+		}
+
+		if !snapshot.HasPaths(targets) {
+			continue
+		}
+
+		latest = snapshot.Time
+		latestID = snapshotID
+		found = true
 	}
 
 	if !found {
@@ -53,7 +63,7 @@ func FindSnapshot(repo Repository, s string) (ID, error) {
 
 // FindFilteredSnapshots yields Snapshots filtered from the list of all
 // snapshots.
-func FindFilteredSnapshots(ctx context.Context, repo Repository, host string, tags []string, paths []string) Snapshots {
+func FindFilteredSnapshots(ctx context.Context, repo Repository, host string, tags []TagList, paths []string) Snapshots {
 	results := make(Snapshots, 0, 20)
 
 	for id := range repo.List(ctx, SnapshotFile) {
@@ -62,7 +72,7 @@ func FindFilteredSnapshots(ctx context.Context, repo Repository, host string, ta
 			fmt.Fprintf(os.Stderr, "could not load snapshot %v: %v\n", id.Str(), err)
 			continue
 		}
-		if (host != "" && host != sn.Hostname) || !sn.HasTags(tags) || !sn.HasPaths(paths) {
+		if (host != "" && host != sn.Hostname) || !sn.HasTagList(tags) || !sn.HasPaths(paths) {
 			continue
 		}